requirements.txt
# API & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Image
Pillow>=10.0.0

# Configuration
pyyaml>=6.0

# Utilities
pydantic>=2.5.0

# 训练时额外安装: pip install torch torchvision numpy pandas
